# crates/mantissa-rest/openapi/mantissa-rest.openapi.json
{
  "components": {
    "schemas": {
...
      "ClusterView": {
        "description": "REST-facing cluster view identifier.",
        "properties": {
          "cluster_id": {
            "type": "string"
          },
          "epoch": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "cluster_id",
          "epoch"
        ],
        "type": "object"
      },
...
      "NodeActionResponse": {
        "description": "REST response returned after a node maintenance action is accepted.",
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "node_id": {
            "type": "string"
          }
        },
        "required": [
          "node_id",
          "accepted"
        ],
        "type": "object"
      },
      "NodeDrainRequest": {
        "additionalProperties": false,
        "description": "REST request body for requesting node drain.",
        "properties": {
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "task_stop_timeout_secs": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "NodeDrainStatus": {
        "description": "REST-facing drain status snapshot for one node.",
        "properties": {
          "blocking_standalone_tasks": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "drain_requested": {
            "type": "boolean"
          },
          "last_scheduling_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "node_id": {
            "type": "string"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "remaining_reserved_gpus": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "remaining_reserved_slots": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "remaining_service_tasks": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "schedulable": {
            "type": "boolean"
          },
          "scheduler_summary_known": {
            "type": "boolean"
          },
          "state": {
            "type": "string"
          },
          "task_stop_timeout_secs": {
            "format": "int32",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "node_id",
          "schedulable",
          "drain_requested",
          "state",
          "remaining_service_tasks",
          "blocking_standalone_tasks",
          "remaining_reserved_slots",
          "remaining_reserved_gpus",
          "scheduler_summary_known",
          "message"
        ],
        "type": "object"
      },
...
      "NodeSummary": {
        "description": "REST-facing node summary returned by topology read routes.",
        "properties": {
          "drain_state": {
            "type": "string"
          },
          "endpoint": {
            "type": "string"
          },
          "health": {
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "readiness": {
            "type": "string"
          },
          "schedulable": {
            "type": "boolean"
          },
          "scheduling_reason": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "hostname",
          "endpoint",
          "health",
          "readiness",
          "schedulable",
          "drain_state",
          "labels"
        ],
        "type": "object"
      },
...
      "RestErrorBody": {
        "description": "Stable JSON error body returned by REST handlers.",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "Mantissa REST token",
        "description": "Daemon-generated REST token presented as a Bearer token.",
        "scheme": "bearer",
        "type": "http"
      },
      "mtls": {
        "description": "Required when the REST listener is bound to a non-loopback interface.",
        "type": "mutualTLS"
      }
    }
  },
  "info": {
    "description": "Local HTTP facade over the Mantissa Cap'n Proto admin session. All /v1 endpoints require an Authorization: Bearer token. Non-loopback listeners must also be served with TLS and client certificate verification enabled.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Mantissa REST API",
    "version": "0.1.0"
  },
  "openapi": "3.2.0",
  "paths": {
...
    "/v1/clusters/current": {
      "get": {
        "operationId": "current",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterView"
                }
              }
            },
            "description": "Current active cluster view."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Returns the active cluster view associated with the local session.",
        "tags": [
          "clusters"
        ]
      }
    },
...
    "/v1/nodes": {
      "get": {
        "operationId": "list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NodeSummary"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Cluster nodes visible to the local daemon."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Lists cluster nodes visible to the local daemon.",
        "tags": [
          "nodes"
        ]
      }
    },
    "/v1/nodes/{node_id}": {
      "delete": {
        "operationId": "evict",
        "parameters": [
          {
            "description": "Node UUID string.",
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeActionResponse"
                }
              }
            },
            "description": "Eviction request accepted."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Evicts one stale node identity by UUID string.",
        "tags": [
          "nodes"
        ]
      },
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "description": "Node UUID string.",
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeSummary"
                }
              }
            },
            "description": "Cluster node summary."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Fetches one cluster node by UUID string.",
        "tags": [
          "nodes"
        ]
      }
    },
    "/v1/nodes/{node_id}/drain": {
      "get": {
        "operationId": "drain_status",
        "parameters": [
          {
            "description": "Node UUID string.",
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDrainStatus"
                }
              }
            },
            "description": "Current node drain status."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Fetches the current drain-status snapshot for one node.",
        "tags": [
          "nodes"
        ]
      },
      "post": {
        "operationId": "drain",
        "parameters": [
          {
            "description": "Node UUID string.",
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeDrainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeActionResponse"
                }
              }
            },
            "description": "Drain request accepted."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Requests drain for one node by UUID string.",
        "tags": [
          "nodes"
        ]
      }
    },
    "/v1/nodes/{node_id}/labels": {
      "put": {
        "operationId": "labels",
        "parameters": [
          {
            "description": "Node UUID string.",
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeLabelsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeLabelsResponse"
                }
              }
            },
            "description": "Label update accepted."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Applies one node label update by UUID string.",
        "tags": [
          "nodes"
        ]
      }
    },
    "/v1/nodes/{node_id}/resume": {
      "post": {
        "operationId": "resume",
        "parameters": [
          {
            "description": "Node UUID string.",
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeActionResponse"
                }
              }
            },
            "description": "Resume request accepted."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Resumes scheduling for one drained node by UUID string.",
        "tags": [
          "nodes"
        ]
      }
    },
    "/v1/scheduler/summary": {
      "get": {
        "operationId": "summary",
        "parameters": [
          {
            "in": "path",
            "name": "peer_id",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "path",
            "name": "details",
            "required": true,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulerSummary"
                }
              }
            },
            "description": "Scheduler capacity summary."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Fetches scheduler capacity summary from the local scheduler capability.",
        "tags": [
          "scheduler"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}